+++ src/main/java/pl/asie/computronics/reference/Names.java
@@ -6,6 +6,11 @@ package pl.asie.computronics.reference;
 public class Names {
 
 	public static final String AE2_SpatialIO = "spatial_io";
+	public static final String EnderIO_CapacitorBank = "capacitor_bank";
+	public static final String EnderIO_IOConfigurable = "io_configurable_tile";
+	public static final String EnderIO_MachineTile = "enderio_machine";
+	public static final String EnderIO_PoweredTile = "powered_tile";
+	public static final String EnderIO_RedstoneTile = "redstone_tile";
 	public static final String FSP_SteamTransporter = "steam_transporter";
 	public static final String FZ_ChargeConductor = "charge_conductor";
 	public static final String MFR_DSU = "deep_storage_unit";
@@ -14,10 +19,8 @@ public class Names {
 	public static final String Railcraft_ElectricGrid = "electric_tile";
 	public static final String Railcraft_LimiterTrack = "limiter_track";
 	public static final String Railcraft_LocomotiveTrack = "locomotive_track";
-	public static final String Railcraft_ReceiverBox = "receiver_box";
 	public static final String Railcraft_RoutingDetector = "routing_detector";
 	public static final String Railcraft_RoutingSwitch = "routing_switch";
 	public static final String Railcraft_RoutingTrack = "routing_track";
 	public static final String RedLogic_Lamp = "lamp";
-
 }